refactor: abstract lookup client creation using factory pattern
