+++ lmcache/integration/vllm/vllm_v1_adapter.py
@@ -15,7 +15,6 @@
 # Standard
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any, Optional
-import threading
 
 # Third Party
 from vllm.config import VllmConfig
@@ -24,12 +23,9 @@ from vllm.distributed.kv_transfer.kv_connector.v1.base import (
     KVConnectorMetadata,
     KVConnectorRole,
 )
-from vllm.utils import cdiv, make_zmq_socket
+from vllm.utils import cdiv
 from vllm.v1.core.sched.output import SchedulerOutput
-from vllm.v1.serial_utils import MsgpackDecoder, MsgpackEncoder
 import torch
-import vllm.envs as envs
-import zmq
 
 # First Party
 from lmcache.integration.vllm.utils import (
@@ -40,8 +36,8 @@ from lmcache.integration.vllm.utils import (
 from lmcache.integration.vllm.vllm_adapter import init_lmcache_engine
 from lmcache.logging import init_logger
 from lmcache.utils import _lmcache_nvtx_annotate
-from lmcache.v1.cache_engine import LMCacheEngine
 from lmcache.v1.compute.blend import LMCBlenderBuilder
+from lmcache.v1.lookup_client import LookupClientFactory
 
 if TYPE_CHECKING:
     # Third Party
@@ -53,89 +49,6 @@ if TYPE_CHECKING:
     from vllm.v1.request import Request
 
 logger = init_logger(__name__)
-
-
-def get_zmq_rpc_path_lmcache(
-    role: KVConnectorRole,
-    is_tp: bool = False,
-    vllm_config: Optional["VllmConfig"] = None,
-) -> str:
-    base_url = envs.VLLM_RPC_BASE_PATH
-    # Default to 0 if not configured
-    rpc_port = 0
-    if vllm_config is not None:
-        rpc_port = vllm_config.kv_transfer_config.get_from_extra_config(
-            "lmcache_rpc_port", 0
-        )
-    logger.debug("Base URL: %s, RPC Port: %s", base_url, rpc_port)
-    return f"ipc://{base_url}/lmcache_rpc_port_{rpc_port}"
-
-
-# TODO: move this to LMCache so that we can gracefully close it
-class LMCacheLookupClient:
-    def __init__(self, role: KVConnectorRole, is_tp: bool, vllm_config: "VllmConfig"):
-        self.encoder = MsgpackEncoder()
-        self.ctx = zmq.Context()  # type: ignore[attr-defined]
-        socket_path = get_zmq_rpc_path_lmcache(role, is_tp, vllm_config)
-        self.socket = make_zmq_socket(
-            self.ctx,
-            socket_path,
-            zmq.REQ,  # type: ignore[attr-defined]
-            bind=False,
-        )
-
-    def lookup(self, token_ids: torch.Tensor) -> int:
-        request = self.encoder.encode(token_ids)
-        self.socket.send_multipart(request, copy=False)
-        resp = self.socket.recv()
-        result = int.from_bytes(resp, "big")
-        return result
-
-    def close(self):
-        self.socket.close(linger=0)
-
-
-class LMCacheLookupServer:
-    def __init__(
-        self,
-        lmcache_engine: LMCacheEngine,
-        role: KVConnectorRole,
-        is_tp: bool,
-        vllm_config: "VllmConfig",
-    ):
-        self.decoder = MsgpackDecoder(torch.Tensor)
-        self.ctx = zmq.Context()  # type: ignore[attr-defined]
-        socket_path = get_zmq_rpc_path_lmcache(role, is_tp, vllm_config)
-        self.socket = make_zmq_socket(
-            self.ctx,
-            socket_path,
-            zmq.REP,  # type: ignore[attr-defined]
-            bind=True,
-        )
-
-        self.lmcache_engine = lmcache_engine
-        self.running = True
-
-        def process_request():
-            while self.running:
-                # try:
-                # request = self.socket.recv()
-                frames = self.socket.recv_multipart(copy=False)
-                token_ids = self.decoder.decode(frames)
-                result = self.lmcache_engine.lookup(token_ids, pin=True)
-                response = result.to_bytes(4, "big")
-                self.socket.send(response)
-                # except Exception as e:
-                #    logger.error("Error in LMCache lookup server: %s", e)
-                #    break
-                # continue
-
-        self.thread = threading.Thread(target=process_request, daemon=True)
-        self.thread.start()
-
-    def close(self):
-        self.socket.close(linger=0)
-        # TODO: close the thread!
 
 
 @dataclass
@@ -394,18 +307,10 @@ class LMCacheConnectorV1Impl:
         config = lmcache_get_config()
         self.layerwise_retrievers = []
         if role == KVConnectorRole.SCHEDULER:
-            # Check if mooncake_lookup_client is configured
-            if config.mooncake_lookup_client is not None:
-                # First Party
-                from lmcache.v1.lookup_client.mooncake_lookup_client import (
-                    MooncakeLookupClient,
-                )
-
-                self.lookup_client = MooncakeLookupClient(
-                    role, is_tp, vllm_config, config.mooncake_lookup_client
-                )
-            else:
-                self.lookup_client = LMCacheLookupClient(role, is_tp, vllm_config)
+            # Create lookup client using factory
+            self.lookup_client = LookupClientFactory.create_lookup_client(
+                role, is_tp, vllm_config
+            )
             self._requests_in_step: dict[str, Request] = {}
         else:
             self.lmcache_engine = init_lmcache_engine(
@@ -425,16 +330,11 @@ class LMCacheConnectorV1Impl:
                     self.lmcache_engine.gpu_connector,
                 )
 
-            # NOTE: Only create the KV lookup API server on worker rank 0
-            # when there are multiple workers
+            # Create lookup server using factory
             assert self.lmcache_engine is not None
-            if (
-                vllm_config.parallel_config.rank == 0
-                and config.mooncake_lookup_client is None
-            ):
-                self.lookup_server = LMCacheLookupServer(
-                    self.lmcache_engine, role, is_tp, vllm_config
-                )
+            self.lookup_server = LookupClientFactory.create_lookup_server(
+                self.lmcache_engine, role, is_tp, vllm_config
+            )
 
         self.kv_caches: dict[str, torch.Tensor] = {}
