+++ lmcache/v1/lookup_client/abstract_client.py
@@ -1,0 +1,55 @@
+# Copyright 2024-2025 LMCache Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# Standard
+from typing import TYPE_CHECKING
+import abc
+
+# Third Party
+import torch
+
+if TYPE_CHECKING:
+    # Third Party
+    pass
+
+
+class LookupClientInterface(metaclass=abc.ABCMeta):
+    """Abstract interface for lookup clients."""
+
+    @abc.abstractmethod
+    def lookup(self, token_ids: torch.Tensor) -> int:
+        """
+        Perform lookup for the given token IDs.
+
+        Args:
+            token_ids: The token IDs to lookup
+
+        Returns:
+            The number of tokens that can be loaded from cache
+        """
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def close(self) -> None:
+        """Close the lookup client and clean up resources."""
+        raise NotImplementedError
+
+    def supports_producer_reuse(self) -> bool:
+        """
+        Return whether this lookup client supports producer KV cache reuse.
+
+        Returns:
+            True if producer reuse is supported, False otherwise
+        """
+        return False
